+++ src/sidebar/trees/items/ProjectTreeItem.ts
@@ -3,7 +3,6 @@
 import { join } from 'path';
 import { TreeItem } from 'vscode';
 
-import { Slot } from '../../../@types/hotkeys';
 import { Project } from '../../../@types/workspaces';
 
 //	Variables __________________________________________________________________
@@ -24,22 +23,18 @@ export class ProjectTreeItem extends TreeItem {
 		title: 'Open Project',
 	};
 	
-	public constructor (public readonly project:Project, public readonly slot:Slot|null, isSubProject:boolean = false) {
+	public constructor (public readonly project:Project, info:string, isSubProject:boolean = false) {
 		
 		super(project.label);
 		
 		const type = project.type;
-		const info:string[] = [];
 		let icon = `${type}`;
-		
-		if (slot) info.push(`[${slot.index}]`);
-		if (project.deleted) info.push('Path does not exist');
 		
 		if (type === 'folder' || type === 'folders') icon += `-color-${project.color || 0}`;
 		
 		this.contextValue = `${isSubProject ? 'sub' : ''}project-${type}`;
 		this.tooltip = project.path;
-		this.description = info.join(' ');
+		this.description = info;
 		
 		this.iconPath = {
 			light: join(basePath, `project-${icon}-light.svg`),